# PackageFluxApp-admin/docs/swagger.yaml
basePath: /api
definitions:
...
  domain.UserInfo:
    properties:
      avatar:
        description: 头像
        type: string
      belongDept:
        description: 数据归属部门
        type: integer
      createTime:
        description: 创建时间
        type: string
      creator:
        description: 创建人
        type: integer
      email:
        description: 邮箱
        type: string
      id:
        description: 主键ID
        type: integer
      modifier:
        description: 修改人
        type: integer
      recordId:
        description: 记录ID
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示排序
        type: integer
      status:
        description: 状态
        type: boolean
      total:
        description: 接口总次数
        type: integer
      updateTime:
        description: 更新时间
        type: string
      usageNumber:
        description: 接口调用次数
        type: integer
      username:
        description: 用户名
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
      success:
        description: 是否成功
        type: boolean
    type: object
...
  web.LoginRequest:
    properties:
      code:
        description: 验证码
        maxLength: 6
        minLength: 6
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        maxLength: 16
        minLength: 6
        type: string
    required:
    - code
    - email
    - password
    type: object
  web.RegisterRequest:
    properties:
      code:
        description: 验证码
        maxLength: 6
        minLength: 6
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        maxLength: 16
        minLength: 6
        type: string
      username:
        description: 用户名
        maxLength: 20
        minLength: 3
        type: string
    required:
    - code
    - email
    - password
    - username
    type: object
  web.SendLoginRequest:
    properties:
      email:
        description: 邮箱
        type: string
    required:
    - email
    type: object
  web.SendRegisterRequest:
    properties:
      email:
        description: 邮箱
        type: string
    required:
    - email
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: 2224693191@qq.com
    name: CJiaの用心
    url: http://www.swagger.io/support
  description: PackageFluxApp-admin在线接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: PackageFluxApp-admin
  version: "1.0"
paths:
...
  /v1/auth/email-login:
    post:
      consumes:
      - application/json
      description: 邮箱验证码登录
      operationId: EmailLogin
      parameters:
      - description: 登录参数
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/web.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 邮箱验证码登录
      tags:
      - 用户登录
  /v1/auth/email-register:
    post:
      consumes:
      - application/json
      description: 邮箱验证码注册
      operationId: EmailRegister
      parameters:
      - description: 注册参数
        in: body
        name: RegisterRequest
        required: true
        schema:
          $ref: '#/definitions/web.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 邮箱验证码注册
      tags:
      - 用户注册
  /v1/auth/send-login-captcha:
    post:
      consumes:
      - application/json
      description: 发送邮箱登录验证码
      operationId: SendEmailLoginCaptcha
      parameters:
      - description: 验证码参数
        in: body
        name: SendLoginRequest
        required: true
        schema:
          $ref: '#/definitions/web.SendLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送邮箱登录验证码
      tags:
      - 用户登录
  /v1/auth/send-register-captcha:
    post:
      consumes:
      - application/json
      description: 发送邮箱注册验证码
      operationId: SendEmailRegisterCaptcha
      parameters:
      - description: 验证码参数
        in: body
        name: SendRegisterRequest
        required: true
        schema:
          $ref: '#/definitions/web.SendRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送邮箱注册验证码
      tags:
      - 用户注册
  /v1/auth/userinfo:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      operationId: UserInfo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 用户登录
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"